refacotring 4
