+++ tools/accuracy_checker/accuracy_checker/postprocessor/resize_segmentation_mask.py
@@ -16,12 +16,12 @@ limitations under the License.
 
 from functools import singledispatch
 import numpy as np
+from PIL import Image
 
 from ..config import NumberField
 from ..utils import get_size_from_config
 from .postprocessor import PostprocessorWithSpecificTargets
 from ..representation import SegmentationPrediction, SegmentationAnnotation
-from ..data_readers import ScipyImageReader
 
 
 class ResizeSegmentationMask(PostprocessorWithSpecificTargets):
@@ -62,7 +62,7 @@ class ResizeSegmentationMask(PostprocessorWithSpecificTargets):
         def _(entry, height, width):
             entry_mask = []
             for class_mask in entry.mask:
-                resized_mask = ScipyImageReader.imresize(class_mask, (height, width))
+                resized_mask = self.resize(class_mask, width, height)
                 entry_mask.append(resized_mask)
             entry.mask = np.array(entry_mask)
 
@@ -70,7 +70,7 @@ class ResizeSegmentationMask(PostprocessorWithSpecificTargets):
 
         @resize_segmentation_mask.register(SegmentationAnnotation)
         def _(entry, height, width):
-            entry.mask = ScipyImageReader.imresize(entry.mask, (height, width))
+            entry.mask = self.resize(entry.mask, width, height)
             return entry
 
         for target in annotation:
@@ -80,3 +80,52 @@ class ResizeSegmentationMask(PostprocessorWithSpecificTargets):
             resize_segmentation_mask(target, target_height, target_width)
 
         return annotation, prediction
+
+    def to_image(self, arr):
+        data = np.asarray(arr)
+        if np.iscomplexobj(data):
+            raise ValueError("Cannot convert a complex-valued array.")
+        shape = list(data.shape)
+        if len(shape) == 2:
+            return self._process_2d(data, shape)
+        if len(shape) == 3 and ((3 in shape) or (4 in shape)):
+            return self._process_3d(data, shape)
+        raise ValueError("'arr' does not have a suitable array shape for any mode.")
+
+    def _process_2d(self, data, shape):
+        shape = (shape[1], shape[0])  # columns show up first
+        bytedata = self._bytescale(data)
+        image = Image.frombytes('L', shape, bytedata.tostring())
+
+        return image
+
+    def _process_3d(self, data, shape):
+        # if here then 3-d array with a 3 or a 4 in the shape length.
+        # Check for 3 in datacube shape --- 'RGB' or 'YCbCr'
+        bytedata = self._bytescale(data)
+        height, width, channels = shape
+        mode = 'RGB' if channels == 3 else 'RGBA'
+        image = Image.frombytes(mode, (width, height), bytedata.tostring())
+
+        return image
+
+    @staticmethod
+    def _bytescale(data):
+        if data.dtype == np.uint8:
+            return data
+        cmin = data.min()
+        cmax = data.max()
+        cscale = cmax - cmin
+        if cscale == 0:
+            cscale = 1
+
+        scale = float(255) / cscale
+        bytedata = (data - cmin) * scale
+
+        return (bytedata.clip(0, 255) + 0.5).astype(np.uint8)
+
+    def resize(self, mask, width, height):
+        image = self.to_image(mask)
+        image_new = image.resize((width, height), resample=0)
+
+        return np.array(image_new)
